[refactor]: UiState 파일 위치 이동 (#65)
