+++ app/src/main/java/com/texthip/thip/ui/group/done/viewmodel/GroupDoneUiState.kt
@@ -1,15 +1,15 @@
-package com.texthip.thip.ui.group.myroom.mock
+package com.texthip.thip.ui.group.done.viewmodel
 
 import com.texthip.thip.ui.group.done.mock.MyRoomCardData
 
-data class GroupMyUiState(
-    val myRooms: List<MyRoomCardData> = emptyList(),
-    val currentRoomType: RoomType = RoomType.PLAYING_AND_RECRUITING,
+data class GroupDoneUiState(
+    val expiredRooms: List<MyRoomCardData> = emptyList(),
     val isLoading: Boolean = false,
     val isLoadingMore: Boolean = false,
     val hasMore: Boolean = true,
+    val userName: String = "",
     val error: String? = null
 ) {
-    val hasContent: Boolean get() = myRooms.isNotEmpty()
+    val hasContent: Boolean get() = expiredRooms.isNotEmpty()
     val canLoadMore: Boolean get() = !isLoading && !isLoadingMore && hasMore
 }
