+++ app/src/main/java/com/texthip/thip/ui/group/myroom/viewmodel/GroupMyUiState.kt
@@ -1,0 +1,16 @@
+package com.texthip.thip.ui.group.myroom.viewmodel
+
+import com.texthip.thip.ui.group.done.mock.MyRoomCardData
+import com.texthip.thip.ui.group.myroom.mock.RoomType
+
+data class GroupMyUiState(
+    val myRooms: List<MyRoomCardData> = emptyList(),
+    val currentRoomType: RoomType = RoomType.PLAYING_AND_RECRUITING,
+    val isLoading: Boolean = false,
+    val isLoadingMore: Boolean = false,
+    val hasMore: Boolean = true,
+    val error: String? = null
+) {
+    val hasContent: Boolean get() = myRooms.isNotEmpty()
+    val canLoadMore: Boolean get() = !isLoading && !isLoadingMore && hasMore
+}
